+++ src/plugins/ncr/components/create-node-modal/index.jsx
@@ -79,8 +79,6 @@ export default class CreateNodeModal extends React.Component {
         onBlurSlug={this.handleBlurSlug}
         onReset={delegate.handleReset}
         onSubmit={delegate.handleSubmit}
-        validate={delegate.handleValidate}
-        warn={delegate.handleWarn}
         {...formConfigs}
       />
     );