+++ src/plugins/ncr/components/slug-editor/index.jsx
@@ -101,6 +101,7 @@ class SlugEditor extends React.Component {
   handleBlur() {
     const { delegate, formName } = this.props;
     const { slug } = this.state;
+
     this.setState(() => ({
       slug: normalizeUnfinishedSlug(slug, slug, delegate.getSluggableConfig(formName)),
     }), this.validateSlug);
@@ -110,7 +111,7 @@ class SlugEditor extends React.Component {
     this.setState((prevState, { initialSlug }) => ({
       slug,
       touched: slug !== initialSlug,
-    }), this.validateSlug);
+    }));
   }
 
   addDate() {
@@ -123,12 +124,17 @@ class SlugEditor extends React.Component {
         slug: newSlug,
         touched: newSlug !== initialSlug,
       };
-    }, this.validateSlug);
+    });
   }
 
   handleRename() {
     const { delegate, initialSlug } = this.props;
     const { slug } = this.state;
+
+    if (slug === initialSlug) {
+      this.handleToggleModal();
+      return;
+    }
 
     if (slug) {
       delegate.handleRename(initialSlug, slug);